Refactors ferry packages
