+++ app/src/main/java/gov/wa/wsdot/android/wsdot/di/FragmentModule.java
@@ -8,12 +8,12 @@ import gov.wa.wsdot.android.wsdot.ui.amtrakcascades.AmtrakCascadesSchedulesDetai
 import gov.wa.wsdot.android.wsdot.ui.borderwait.BorderWaitNorthboundFragment;
 import gov.wa.wsdot.android.wsdot.ui.borderwait.BorderWaitSouthboundFragment;
 import gov.wa.wsdot.android.wsdot.ui.camera.CameraImageFragment;
-import gov.wa.wsdot.android.wsdot.ui.ferries.schedules.bulletins.FerriesRouteAlertsBulletinDetailsFragment;
-import gov.wa.wsdot.android.wsdot.ui.ferries.schedules.bulletins.FerriesRouteAlertsBulletinsFragment;
-import gov.wa.wsdot.android.wsdot.ui.ferries.schedules.sailings.departures.FerriesRouteSchedulesDayDeparturesFragment;
-import gov.wa.wsdot.android.wsdot.ui.ferries.schedules.sailings.departures.FerriesTerminalCameraFragment;
-import gov.wa.wsdot.android.wsdot.ui.ferries.schedules.FerriesRouteSchedulesFragment;
-import gov.wa.wsdot.android.wsdot.ui.ferries.schedules.sailings.departures.VesselWatchFragment;
+import gov.wa.wsdot.android.wsdot.ui.ferries.bulletins.FerriesRouteAlertsBulletinDetailsFragment;
+import gov.wa.wsdot.android.wsdot.ui.ferries.bulletins.FerriesRouteAlertsBulletinsFragment;
+import gov.wa.wsdot.android.wsdot.ui.ferries.departures.FerriesRouteSchedulesDayDeparturesFragment;
+import gov.wa.wsdot.android.wsdot.ui.ferries.departures.FerriesTerminalCameraFragment;
+import gov.wa.wsdot.android.wsdot.ui.ferries.FerriesRouteSchedulesFragment;
+import gov.wa.wsdot.android.wsdot.ui.ferries.departures.vesselwatch.VesselWatchFragment;
 import gov.wa.wsdot.android.wsdot.ui.home.DashboardFragment;
 import gov.wa.wsdot.android.wsdot.ui.home.FavoritesFragment;
 import gov.wa.wsdot.android.wsdot.ui.home.HighImpactAlertsFragment;
