+++ app/src/main/java/gov/wa/wsdot/android/wsdot/di/ViewModelModule.java
@@ -13,10 +13,10 @@ import gov.wa.wsdot.android.wsdot.ui.borderwait.BorderWaitViewModel;
 import gov.wa.wsdot.android.wsdot.ui.camera.CameraViewModel;
 import gov.wa.wsdot.android.wsdot.ui.camera.MapCameraViewModel;
 import gov.wa.wsdot.android.wsdot.ui.ferries.FerrySchedulesViewModel;
-import gov.wa.wsdot.android.wsdot.ui.ferries.schedules.bulletins.FerriesBulletinsViewModel;
-import gov.wa.wsdot.android.wsdot.ui.ferries.schedules.sailings.departures.FerryTerminalCameraViewModel;
-import gov.wa.wsdot.android.wsdot.ui.ferries.schedules.sailings.departures.FerryTerminalViewModel;
-import gov.wa.wsdot.android.wsdot.ui.ferries.vesselwatch.VesselWatchViewModel;
+import gov.wa.wsdot.android.wsdot.ui.ferries.bulletins.FerriesBulletinsViewModel;
+import gov.wa.wsdot.android.wsdot.ui.ferries.departures.FerryTerminalCameraViewModel;
+import gov.wa.wsdot.android.wsdot.ui.ferries.departures.FerryTerminalViewModel;
+import gov.wa.wsdot.android.wsdot.ui.ferries.departures.vesselwatch.VesselWatchViewModel;
 import gov.wa.wsdot.android.wsdot.ui.home.FavoritesViewModel;
 import gov.wa.wsdot.android.wsdot.ui.mountainpasses.MountainPassViewModel;
 import gov.wa.wsdot.android.wsdot.ui.myroute.MyRouteViewModel;
