+++ app/src/main/java/gov/wa/wsdot/android/wsdot/ui/ferries/FerriesRouteSchedulesActivity.java
@@ -16,7 +16,7 @@
  *
  */
 
-package gov.wa.wsdot.android.wsdot.ui.ferries.schedules;
+package gov.wa.wsdot.android.wsdot.ui.ferries;
 
 import android.content.Intent;
 import android.net.Uri;
