+++ app/src/main/java/gov/wa/wsdot/android/wsdot/ui/ferries/bulletins/FerriesBulletinsViewModel.java
@@ -1,4 +1,4 @@
-package gov.wa.wsdot.android.wsdot.ui.ferries.schedules.bulletins;
+package gov.wa.wsdot.android.wsdot.ui.ferries.bulletins;
 
 import android.arch.lifecycle.LiveData;
 import android.arch.lifecycle.MutableLiveData;
