+++ app/src/main/java/gov/wa/wsdot/android/wsdot/ui/ferries/bulletins/FerriesRouteAlertsBulletinsFragment.java
@@ -16,7 +16,7 @@
  *
  */
 
-package gov.wa.wsdot.android.wsdot.ui.ferries.schedules.bulletins;
+package gov.wa.wsdot.android.wsdot.ui.ferries.bulletins;
 
 import android.arch.lifecycle.ViewModelProvider;
 import android.arch.lifecycle.ViewModelProviders;
